Derive Macro diesel_derives::Associations

#[derive(Associations)]
{
    // Attributes available to this derive:
    #[belongs_to]
    #[column_name]
    #[table_name]
}

Implement required traits for the associations API

This derive implement support for diesels associations api. Check the module level documentation of the diesel::associations module for details.

Attributes

Required type attributes

#[belongs_to(User)], to specify a child-to-parent relation ship between the current type and the specified parent type (User). If this attribute is given multiple times, multiple relation ships are generated.

#[belongs_to(User, foreign_key = "mykey")], variant of the attribute above. Allows to specify the name of the foreign key. If the foreign key is not specified explicitly, the remote lower case type name with an appended _id is used as foreign key name. (user_id in this example case)

Optional type attributes

#[table_name = "some_table_name"] specifies the table this type belongs to. Requires that some_table_name is in scope. If this attribute is not used, the type name converted to snake_case with an added s is used as table name

Optional field attributes

#[column_name = "some_table_name"], overrides the column the current field maps to to some_table_name. By default the field name is used as column name. Only useful for the foreign key field.