Derive Macro diesel_derives::Insertable

#[derive(Insertable)]
{
    // Attributes available to this derive:
    #[table_name]
    #[column_name]
    #[diesel]
}

Implements Insertable

To implement Insertable this derive needs to know the corresponding table type. By default it uses the snake_case type name with an added s. It is possible to change this default by using #[table_name = "something"]. In both cases the module for that table must be in scope. For example, to derive this for a struct called User, you will likely need a line such as use schema::users;

If a field name of your struct differs from the name of the corresponding column, you can annotate the field with #[column_name = "some_column_name"].

Your struct can also contain fields which implement Insertable. This is useful when you want to have one field map to more than one column (for example, an enum that maps to a label and a value column). Add #[diesel(embed)] to any such fields.

To provide custom serialization behavior for a field, you can use #[diesel(serialize_as = "SomeType")]. If this attribute is present, Diesel will call .into on the corresponding field and serialize the instance of SomeType, rather than the actual field on your struct. This can be used to add custom behavior for a single field, or use types that are otherwise unsupported by Diesel. Using #[diesel(serialize_as)] is incompatible with #[diesel(embed)]. Normally, Diesel produces two implementations of the Insertable trait for your struct using this derive: one for an owned version and one for a borrowed version. Using #[diesel(serialize_as)] implies a conversion using .into which consumes the underlying value. Hence, once you use #[diesel(serialize_as)], Diesel can no longer insert borrowed versions of your struct.

Attributes

Optional type attributes

#[table_name = "some_table_name"], specifies the table this type is insertable into. Requires that some_table_name is in scope. If this attribute is not used, the type name converted to snake_case with an added s is used as table name

Optional field attributes

#[column_name = "some_table_name"], overrides the column the current field maps to some_table_name. By default the field name is used as column name

#[diesel(embed)], specifies that the current field maps not only to single database field, but is a struct that implements Insertable

#[diesel(serialize_as = "SomeType")], instead of serializing the actual field type, Diesel will convert the field into SomeType using .into and serialize that instead. By default this derive will serialize directly using the actual field type.

Examples

If we want to customize the serialization during insert, we can use #[diesel(serialize_as)].

#[derive(Debug, FromSqlRow, AsExpression)]
#[sql_type = "sql_types::Text"]
struct UppercaseString(pub String);

impl Into<UppercaseString> for String {
    fn into(self) -> UppercaseString {
        UppercaseString(self.to_uppercase())
    }
}

impl<DB> ToSql<sql_types::Text, DB> for UppercaseString
    where
        DB: Backend,
        String: ToSql<sql_types::Text, DB>,
{
    fn to_sql<W: Write>(&self, out: &mut Output<W, DB>) -> serialize::Result {
        self.0.to_sql(out)
    }
}

#[derive(Insertable, PartialEq, Debug)]
#[table_name = "users"]
struct InsertableUser {
    id: i32,
    #[diesel(serialize_as = "UppercaseString")]
    name: String,
}

let user = InsertableUser {
    id: 1,
    name: "thomas".to_string(),
};

diesel::insert_into(users)
    .values(user)
    .execute(&connection)
    .unwrap();

assert_eq!(
    Ok("THOMAS".to_string()),
    users.select(name).first(&connection)
);