Crate diesel_derives

Macros

define_sql_function

Declare a sql function for use in your code.

sql_function_procDeprecated

A legacy version of define_sql_function!.

table_proc

Specifies that a table exists, and what columns it has. This will create a new public module, with the same name, as the name of the table. In this module, you will find a unit struct named table, and a unit struct with the name of each column.

Attribute Macros

auto_type

Automatically annotates return type of a query fragment function

Derive Macros

AsChangeset

Implements AsChangeset

AsExpression

Implements all required variants of AsExpression

Associations

Implement required traits for the associations API

DieselNumericOps

Implement numeric operators for the current query node

FromSqlRow

Implements Queryable for types that correspond to a single SQL type. The type must implement FromSql.

Identifiable

Implements Identifiable for references of the current type

Insertable

Implements Insertable

MultiConnection

This derives implements diesel::Connection and related traits for an enum of connections to different databases.

QueryId

Implements QueryId

Queryable

Implements Queryable to load the result of statically typed queries

QueryableByName

Implements QueryableByName for untyped sql queries, such as that one generated by sql_query

Selectable

Implements Selectable

SqlType

Implement necessary traits for adding a new sql type

ValidGrouping

Implements ValidGrouping